refactor: canvas elements
-Refactor all external canvas and data elements into their respective folders.
